Extract method to trait
https://trello.com/c/tcIFtuRl
